avcodec/rkmpp : Fix broken build due to missing control operation
This patch is taking care of https://trac.ffmpeg.org/ticket/6834.
It seems that one of the control operations that was available to get
the free decoders input slots was removed.

There is another control operation to retrieve the used slots. Given
that the input slot count is hardcoded to 4 in mpp at this point,
replacing the old control operation by the other one.

This was tested on Rockchip ROCK64.

Signed-off-by: wm4 <nfxjfg@googlemail.com>
(cherry picked from commit c6f84106366c6f243a8b07dbffcc7880009aa904)

Conflicts:
      configure

Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>